+++ src/common/interfaces/empty.ts
