+++ src/common/interfaces/local-storage-object.ts
