+++ src/components/About/style.scss
@@ -1,4 +1,4 @@
-@import 'mixins.scss';
+@import '../../common/scss/mixins.scss';
 
 .avatar {
   padding-bottom: 1.5rem;
