+++ src/components/LightYoutube/style.scss
@@ -1,5 +1,5 @@
-@import 'variables.scss';
-@import 'mixins.scss';
+@import '../../common/scss/mixins.scss';
+@import '../../common/scss/variables.scss';
 
 .light-youtube {
   @include display-flex();
