+++ src/components/Loader/style.scss
