+++ src/components/ProjectDetails/style.scss
