+++ src/components/ProjectOverview/style.scss
@@ -1,4 +1,4 @@
-@import 'variables.scss';
+@import '../../common/scss/variables.scss';
 
 .clickable:hover {
   cursor: pointer;
